pyproject.toml
[project]
name = "options-data-pipeline"
version = "1.0.0"
description = "Standalone data pipeline for options and OHLC data collection and aggregation"
requires-python = ">=3.11"
readme = "README.md"

dependencies = [
    # Orchestration
    "prefect>=2.14.0",

    # Database
    "psycopg2-binary>=2.9.9",
    "asyncpg>=0.29.0",

    # HTTP Clients
    "httpx>=0.25.0",
    "aiohttp>=3.9.0",
    "requests>=2.31.0",

    # Configuration
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",

    # Utilities
    "pytz>=2023.3",
    "structlog>=23.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
]

[project.scripts]
pipeline = "pipeline.flows.main:main"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.black]
line-length = 100
target-version = ["py311"]
